💡 Fix #5 - 커넥티드 아이디 관련 컨트롤러 주석 추가
