--- src/main/resources/templates/userInfo.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="https://www.thymeleaf.org">
-<head>
-    <title>User Info</title>
-</head>
-<body>
-<h1>User Information</h1>
-<div>
-    <p><strong>Email:</strong> <span th:text="${email}">Email</span></p>
-    <p><strong>Authorities:</strong> <span th:text="${authorities}">Authorities</span></p>
-    <form action="/logout" method="post">
-        <button type="submit">Logout</button>
-    </form>
-</div>
-</body>
-</html>
